server/requirements.txt
dspy-ai
python-dotenv
flask
flask_cors
pydantic
extracty
youtube-transcript-api
openai
pydub